requirements/requirements.txt
opencv-python
h5py
matplotlib
numpy
scipy
tqdm
keras
scikit-learn
pillow
ipykernel
tensorflow